Specify column in_dm_table
The in_dm_field was split into two
to make it more clear which data it depends on.

The column in_dm_table was split into
- pair_mutation_in_dm_table:
Is the compound-target pair in the drug_mechanism table
when taking mutation information into account?
This is identical to the old field in_dm_table.
The drug_mechanism table does not include mutation information,
therefore, all pairs with a mutation are automatically set to False.
This is in line with the way new compound target pairs from
the drug_mechanism table are added to the set of existing pairs
with activity data.
For this, mutation annotations are taken into account.
Therefore, *(cpd A, target B without mutation)* will be added
if a pchembl is present for *(cpd A, target B with mutation C)*
but not for *(cpd A, target B without mutation)*.
This field is also used to assign keep_for_binding
which is used to determine the B subset.
In the B subset of data based on binding asssays,
data without pchembl values is only considered
if it is in the drud_mechanism table
which by default does not include mutation information (see above).

- pair_in_dm_table:
Is the compound-target pair in the drug_mechanism table
when ignoring mutation information?
New field with values sometimes different from pair_mutation_in_dm_table.
Added because the DTI field is assigned without taking mutation
information into account and this field makes more sense in that context.